- Splash updated, Report on timeline added - v1.5.18(13)

+++ SportsYap/View Controllers/Main/Profile/ProfileViewController.swift
@@ -253,12 +253,16 @@ extension ProfileViewController: UITableViewDelegate, UITableViewDataSource {
 
 extension ProfileViewController: ProfilePostTableViewCellDelegate {
     func didTapOption(post: Post) {
-        let alert = UIAlertController(title: NSLocalizedString("Confirm", comment: ""), message: NSLocalizedString("Are you sure you want to delete this post?", comment: ""), preferredStyle: .alert)
-        alert.addAction(UIAlertAction(title: NSLocalizedString("Delete", comment: ""), style: .destructive, handler: { (_) in
-            self.deletePost(post: post)
-        }))
-        alert.addAction(UIAlertAction(title: NSLocalizedString("Cancel", comment: ""), style: .cancel, handler: nil))
-        present(alert, animated: true, completion: nil)
+        if post.user.id == User.me.id {
+            let alert = UIAlertController(title: NSLocalizedString("Confirm", comment: ""), message: NSLocalizedString("Are you sure you want to delete this post?", comment: ""), preferredStyle: .alert)
+            alert.addAction(UIAlertAction(title: NSLocalizedString("Delete", comment: ""), style: .destructive, handler: { (_) in
+                self.deletePost(post: post)
+            }))
+            alert.addAction(UIAlertAction(title: NSLocalizedString("Cancel", comment: ""), style: .cancel, handler: nil))
+            present(alert, animated: true, completion: nil)
+        } else {
+            showAbuseAlert(postId: post.id)
+        }
     }
     
     func didFistBump(post: Post) {
